requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
rasterio>=1.3.0
numpy>=1.26.0
Pillow>=10.0.0
pyproj>=3.6.0
